requirements.txt
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.45.0
accelerate>=1.0.0
peft>=0.13.0
einops>=0.8.0
timm>=1.0.0
datasets>=2.20.0
wandb>=0.17.0
scipy>=1.13.0
matplotlib>=3.9.0
seaborn>=0.13.0
pandas>=2.2.0
pyyaml>=6.0
pillow>=10.0.0
tqdm>=4.66.0
